Separate groups into different cards
